fix manpage on all pages
