+++ src/components/Header/AppHeader.tsx
@@ -15,7 +15,7 @@ const AppHeader = (): JSX.Element => {
   const [menuOpen, setMenuOpen] = useState(false);
   const [loginPopupVisible, setLoginPopupVisible] = useState(false);
   const isLoggedIn = true;
-  const linkElements = routes.map((route) => (
+  const linkElements = routes.filter(route => route.name).map((route) => (
     <NavLink
       key={route.path}
       className="md:text-gray-600 md:hover:text-blue-400 md:dark:text-white p-1 font-semibold text-nowrap"
